refactor tempChat.js and temp.js to rename them to names that make sense.
