+++ public/javascripts/sortSighting.js
